fix: first plugin config
