+++ package.json
@@ -26,7 +26,8 @@
   "dependencies": {
     "@capacitor/camera": "^5.0.9",
     "@peculiar/webcrypto": "^1.4.5",
-    "js-sha3": "^0.9.3"
+    "js-sha3": "^0.9.3",
+    "vue": "^2.7.16"
   },
   "files": [
     "dist"
